+++ changes.md
@@ -20,6 +20,8 @@ This release ...
 
 + Created new function std::vector<double> UniformOnUnitCircle( void ) that returns a (uniformly) random vector (x,y,0) on the unit circle (in the z = 0 plane).  
 
++ Created std::vector<double> LegacyRandomOnUnitSphere() that reproduces old behaviors of creating a random vector on the unit sphere. Never use this except if trying to replicate old results. Always use UniformOnUnitSphere() instead. 
+
 + See 1.7.0 for other recent minor changes.
 
 ### Beta features (not fully supported):